tradingagents-cn/requirements.txt
langgraph>=0.0.55
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-anthropic>=0.1.0

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

pymongo>=4.6.0
redis>=5.0.0

akshare>=1.13.0
tushare>=1.4.0

python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
aiofiles>=23.2.0
aiohttp>=3.9.0
edge-tts>=6.1.0

loguru>=0.7.0
sentry-sdk>=1.40.0
prometheus-fastapi-instrumentator>=6.1.0

pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
